# frontend/openapi.json
{
  "components": {
    "schemas": {
      "Category": {
        "enum": [
          "water",
          "electricity",
          "sanitation",
          "roads",
          "streetlights",
          "other"
        ],
        "title": "Category",
        "type": "string"
      },
      "CategoryCount": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "category",
          "count"
        ],
        "title": "CategoryCount",
        "type": "object"
      },
      "ComplaintCreate": {
        "description": "Validated input payload for creating a new complaint (POST /api/complaints).",
        "properties": {
          "location": {
            "description": "Physical location or area of the issue",
            "examples": [
              "Gulberg III, Main Boulevard, Block C"
            ],
            "maxLength": 200,
            "minLength": 3,
            "title": "Location",
            "type": "string"
          },
          "reporter_contact": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional contact (email, phone) for follow-up",
            "examples": [
              "03001234567"
            ],
            "title": "Reporter Contact"
          },
          "text": {
            "description": "Full description of the municipal complaint",
            "examples": [
              "Bijli nahin aa rahi, transformer khrab ho gaya hai pichle teen din se."
            ],
            "maxLength": 2000,
            "minLength": 10,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "location"
        ],
        "title": "ComplaintCreate",
        "type": "object"
      },
      "ComplaintListResponse": {
        "description": "Paginated list of complaints.",
        "properties": {
          "complaints": {
            "items": {
              "$ref": "#/components/schemas/ComplaintResponse"
            },
            "title": "Complaints",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "complaints",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "ComplaintListResponse",
        "type": "object"
      },
      "ComplaintResponse": {
        "description": "Complaint object returned by GET and POST responses.\n\n``reporter_contact`` is deliberately excluded: the API has no\nauthentication, so returning it would publish every reporter's phone or\nemail on the public list endpoint (see ADR 0004). Enum-typed fields make\nthe allowed values visible in OpenAPI for the typed frontend client.",
        "properties": {
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          },
          "allowed_transitions": {
            "description": "Statuses this complaint may move to next, computed by the backend's state machine. The frontend renders these; it never hardcodes them.",
            "items": {
              "$ref": "#/components/schemas/Status"
            },
            "title": "Allowed Transitions",
            "type": "array"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "triage_corrected_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Timestamp when an operator manually corrected category or priority, if any.",
            "title": "Triage Corrected At"
          },
          "triage_latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triage Latency Ms"
          },
          "triaged_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriagedBy"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "location",
          "category",
          "priority",
          "status",
          "ai_summary",
          "triaged_by",
          "triage_latency_ms",
          "created_at",
          "updated_at",
          "allowed_transitions"
        ],
        "title": "ComplaintResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Priority": {
        "enum": [
          "high",
          "normal",
          "low"
        ],
        "title": "Priority",
        "type": "string"
      },
      "PriorityCount": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          }
        },
        "required": [
          "priority",
          "count"
        ],
        "title": "PriorityCount",
        "type": "object"
      },
      "ProviderMeta": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "fallback_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Provider"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "timeout_seconds": {
            "title": "Timeout Seconds",
            "type": "number"
          },
          "triaged_by": {
            "title": "Triaged By",
            "type": "string"
          }
        },
        "required": [
          "name",
          "triaged_by",
          "active",
          "timeout_seconds",
          "description"
        ],
        "title": "ProviderMeta",
        "type": "object"
      },
      "ProvidersResponse": {
        "properties": {
          "active_provider": {
            "title": "Active Provider",
            "type": "string"
          },
          "cache": {
            "$ref": "#/components/schemas/TriageCacheStats"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderMeta"
            },
            "title": "Providers",
            "type": "array"
          },
          "recent_outcomes": {
            "items": {
              "$ref": "#/components/schemas/TriageOutcome"
            },
            "title": "Recent Outcomes",
            "type": "array"
          }
        },
        "required": [
          "active_provider",
          "providers",
          "recent_outcomes",
          "cache"
        ],
        "title": "ProvidersResponse",
        "type": "object"
      },
      "StatsResponse": {
        "description": "Aggregated statistics for GET /api/stats.",
        "properties": {
          "by_category": {
            "items": {
              "$ref": "#/components/schemas/CategoryCount"
            },
            "title": "By Category",
            "type": "array"
          },
          "by_priority": {
            "items": {
              "$ref": "#/components/schemas/PriorityCount"
            },
            "title": "By Priority",
            "type": "array"
          },
          "in_progress_count": {
            "title": "In Progress Count",
            "type": "integer"
          },
          "open_count": {
            "title": "Open Count",
            "type": "integer"
          },
          "rejected_count": {
            "title": "Rejected Count",
            "type": "integer"
          },
          "resolved_count": {
            "title": "Resolved Count",
            "type": "integer"
          },
          "total_complaints": {
            "title": "Total Complaints",
            "type": "integer"
          }
        },
        "required": [
          "total_complaints",
          "by_category",
          "by_priority",
          "open_count",
          "in_progress_count",
          "resolved_count",
          "rejected_count"
        ],
        "title": "StatsResponse",
        "type": "object"
      },
      "Status": {
        "enum": [
          "open",
          "in_progress",
          "resolved",
          "rejected"
        ],
        "title": "Status",
        "type": "string"
      },
      "StatusUpdateRequest": {
        "description": "Request body for PATCH /api/complaints/{id}/status.",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/Status",
            "description": "Target status to transition to"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusUpdateRequest",
        "type": "object"
      },
      "TriageCacheStats": {
        "description": "AI triage cache effectiveness, counted in Redis across all replicas.",
        "properties": {
          "hit_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hit Rate"
          },
          "hits": {
            "title": "Hits",
            "type": "integer"
          },
          "misses": {
            "title": "Misses",
            "type": "integer"
          }
        },
        "required": [
          "hits",
          "misses",
          "hit_rate"
        ],
        "title": "TriageCacheStats",
        "type": "object"
      },
      "TriageCorrectionRequest": {
        "description": "Request body for PATCH /api/complaints/{id}/triage.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Category"
              },
              {
                "type": "null"
              }
            ],
            "description": "Corrected category"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Priority"
              },
              {
                "type": "null"
              }
            ],
            "description": "Corrected priority"
          }
        },
        "title": "TriageCorrectionRequest",
        "type": "object"
      },
      "TriageOutcome": {
        "description": "One recent triage: which provider answered, how long it took, and whether it fell back.",
        "properties": {
          "complaint_id": {
            "format": "uuid",
            "title": "Complaint Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fallback": {
            "title": "Fallback",
            "type": "boolean"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "complaint_id",
          "provider",
          "latency_ms",
          "fallback",
          "created_at"
        ],
        "title": "TriageOutcome",
        "type": "object"
      },
      "TriagedBy": {
        "enum": [
          "llm:groq",
          "llm:ollama",
          "rules",
          "rules:fallback",
          "simulated"
        ],
        "title": "TriagedBy",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "items": {
              "type": "object"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "CivicPulse: Municipal Complaint Intake, AI Triage, and Operations Platform.\n\nFeatures:\n- Multi-provider AI triage (Groq, Ollama, Rules, Simulated)\n- Resilient fallback architecture\n- Distributed Redis caching and rate-limiting\n- Production observability with Prometheus metrics and structured JSON logging",
    "title": "CivicPulse API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/complaints": {
      "get": {
        "operationId": "list_complaints_api_complaints_get",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Category"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Priority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Status"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List complaints with filters and pagination",
        "tags": [
          "Complaints"
        ]
      },
      "post": {
        "operationId": "create_complaint_api_complaints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Submit a complaint (validated, rate-limited, triaged, persisted)",
        "tags": [
          "Complaints"
        ]
      }
    },
    "/api/complaints/{complaint_id}": {
      "get": {
        "operationId": "get_complaint_api_complaints__complaint_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "complaint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Complaint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Complaint not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get one complaint",
        "tags": [
          "Complaints"
        ]
      }
    },
    "/api/complaints/{complaint_id}/status": {
      "patch": {
        "operationId": "change_status_api_complaints__complaint_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "complaint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Complaint Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Complaint not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Transition not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move a complaint through the status state machine",
        "tags": [
          "Complaints"
        ]
      }
    },
    "/api/complaints/{complaint_id}/triage": {
      "patch": {
        "operationId": "correct_triage_api_complaints__complaint_id__triage_patch",
        "parameters": [
          {
            "in": "path",
            "name": "complaint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Complaint Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriageCorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Complaint not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correct a complaint's priority and/or category",
        "tags": [
          "Complaints"
        ]
      }
    },
    "/api/meta/providers": {
      "get": {
        "description": "Active provider, all known providers, and the last 20 triage outcomes\n(provider, latency, fallback y/n).",
        "operationId": "get_providers_api_meta_providers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvidersResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List AI Triage Providers",
        "tags": [
          "Metadata"
        ]
      }
    },
    "/api/stats": {
      "get": {
        "operationId": "get_stats_api_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Aggregate complaint statistics",
        "tags": [
          "Stats"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "Lightweight liveness check.\n\nCRITICAL: Does NOT touch the database or Redis so it never produces false\nnegatives under database load or transient connectivity blips.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness Probe",
        "tags": [
          "Health & Diagnostics"
        ]
      }
    },
    "/metrics": {
      "get": {
        "description": "Export Prometheus metrics.",
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Prometheus Metrics",
        "tags": [
          "Health & Diagnostics"
        ]
      }
    },
    "/ready": {
      "get": {
        "description": "Readiness probe that checks backing infrastructure (PostgreSQL & Redis).\n\nReturns 200 if all dependencies are healthy.\nReturns 503 naming the specific failed dependency if any check fails.\nOnly the exception class is returned; full errors go to the logs, since\ndriver messages can include hostnames, usernames or DSN fragments.",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readiness Probe",
        "tags": [
          "Health & Diagnostics"
        ]
      }
    }
  }
}